Streamline memory model tests
